+++ market/src/test/java/university/market/member/service/AdminMemberServiceTest.java
@@ -19,7 +19,7 @@ import university.market.member.domain.MemberVO;
 import university.market.member.service.dto.request.JoinRequest;
 import university.market.member.service.dto.request.LoginRequest;
 import university.market.member.service.dto.response.LoginResponse;
-import university.market.member.utils.JwtTokenProvider;
+import university.market.member.utils.jwt.JwtTokenProvider;
 
 @SpringBootTest
 @AutoConfigureMockMvc
