+++ market/src/test/java/university/market/member/service/MemberServiceTest.java
@@ -21,7 +21,7 @@ import university.market.member.exception.MemberException;
 import university.market.member.service.dto.request.JoinRequest;
 import university.market.member.service.dto.request.LoginRequest;
 import university.market.member.service.dto.response.LoginResponse;
-import university.market.member.utils.JwtTokenProvider;
+import university.market.member.utils.jwt.JwtTokenProvider;
 import university.market.verify.email.service.EmailVerificationServiceImpl;
 
 @SpringBootTest
